refactor: rename seeder config
Due to the seeder config being a typescript file this changed the
directory structure of the dist output, resulting in inconsistent
entrypoints for both the annotation and lab-server. This caused the
lab-server to have to be started with `node dist/src/main` and the
annotation-server with `node dist/main` due to the `seeder.ormconfig.ts`
file being included in the top-lever dist folder by `tsc` (as it was not
excluded in `tsconfig.json`). Seeing as the seeder is not necessary in a
production setting anyhow, it makes sense to not include this file in
the dist output. Therefore, it's extension has been changed to `.js`,
although it could have equally as well just been excluded in
`tsconfig.json`.