feat(ci): (#211) add health-checks in deployments
the `docker-compose` command in the respecitve `deploy` jobs for the
lab/annotation-server does not check for the liveliness of the
containers. As such, a container can crash without this reflecting
negatively in the CI workflow due to the `docker-compose` command always
exiting with code 0. To fix this, we add a crude health check which
exits with non-0 status when a container has status "Fail" after a
60-second grace period.

Closes Issue #211