refactor: unify caching interface
There seemed to be multiple different approaches to caching medications
with guidelines within the codebase. Either by calling
`_cacheMedications` or `addUnique`.

These basically do the same thing except that one adds a list of
guideline-medications and the other one only adds a single medication
(also syncing possible out-of-date guidelines).
MedicationWithGuideLine(s) and List<MedicationWithGuideLine>(s) are now
both cached using a private helper in the CachedMedications file.

An API was added to the CachedMedications singleton which wraps this
internal helper.